rename header name and add missing include
