+++ 1_4_polar/main.cpp
@@ -3,7 +3,7 @@
 
 #include "color.h"
 #include "vec2.h"
-#include "math.h" // NOLINT
+#include "mymath.h" // NOLINT
 
 vec3 tex(vec2 st) {
     const std::vector<color> col3 = {
