+++ common/mymath.h
